ensures that user creation sets enabled flag by default
