+++ tests/test_views.py
@@ -90,42 +90,6 @@ def test_statistics(_demo):
     assert "total_variants" in res.keys()
 
 
-def test_create_user(_demo):
-    """res -> tuple(flask.wrappers.Response)"""
-    res = create_user()
-    _check_only_available_to_admin(res)
-
-
-def test_create_user_individual(_demo):
-    """res -> tuple(flask.wrappers.Response)"""
-    res = create_user_individual()
-    _check_only_available_to_admin(res)
-
-
-def test_get_user(_demo):
-    """res -> tuple(flask.wrappers.Response)"""
-    res = get_user("whatever_user")
-    _check_only_available_to_admin(res)
-
-
-def test_get_users(_demo):
-    """res -> tuple(flask.wrappers.Response)"""
-    res = get_users()
-    _check_only_available_to_admin(res)
-
-
-def test_delete_user_individual(_demo):
-    """res -> tuple(flask.wrappers.Response)"""
-    res = delete_user_individual()
-    _check_only_available_to_admin(res)
-
-
-def test_enable_user(_demo):
-    """res -> tuple(flask.wrappers.Response)"""
-    res = enable_user("my_user", "true")
-    _check_only_available_to_admin(res)
-
-
 def _check_only_available_to_admin(res):
     assert res[0].status_code == 200
     assert res[0].data == b'{"error":"Admin permissions required to perform this operation"}\n'
